Extern cross-check: unify caller signatures, adjudicate the rest
Mechanical sweep of every extern declaration against its siblings and
definitions. Unified (measured byte-free, MATCH each):
- draw_time_'s stray s16 x in mission_score_screen -> s32.
- spawn_smoke_burst_'s int/s32 variants in JumpControl and
  ProcItemShinsoku -> the definition's (u16, s16, s16).
- set_impact_ex_'s size/color params in Shinsoku/Gosin -> the
  definition's short/long spelling.
- CreateHumanoid's `Humanoid *vcalloc(...)` lie -> void * with an
  explicit cast at the call.
- CdaStop's void * VSyncCallback -> the function-pointer prototype.
- decode_tmd_fast_'s int DivDepth -> u_long (matches its twin).
Adjudicated byte-required (noted inline):
- set_impact_ex_'s u16 tail params must stay s32 at BOTH callers: they
  pass negative rotate constants and the signed view keeps li -30
  (addiu) instead of ori 0xffe2.
- ActivateHumans's s32 PacketUsed (vs EndDrawing's u32) now carries
  its measurement note; DEBUG_CAMERA_SLOTS_'s anonymous void *[] view
  in the initializer cross-referenced to its typed reader.
Already-documented per-TU divergences left alone: ControlScheme,
STAGE_LAYOUT_NUMBER[] (store-forward lever), volatile StageTime.

+++ src/main.exe/ProcItemGosin.c
@@ -69,8 +69,11 @@ typedef union
 
 /* Retail's caller promotes these scalar arguments before the call. */
 extern void set_impact_ex_(VECTOR *pos, GsCOORDINATE2 *super,
-                           s32 start_size, s32 end_size,
-                           s32 start_color, s32 end_color,
+                           short start_size, short end_size,
+                           long start_color, long end_color,
+                           /* s32 tail vs the definition's u16s is measured
+                            * byte-required (same negative-constant lever as
+                            * ProcItemShinsoku.c). */
                            s32 rotate, s32 rotate_speed, s32 time, s32 type);
 extern VECTOR vec_y_n1200_z_400; /* {0,-1200,400} */
